# zapcore: add stopped for Stop method called exactly once