@zchee 2021-06-19
@prashantv changed to stopped bool and add local value for avoid call mutex twice.
Hiw about it?